+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fixed `onTouchEnd()` fired on `touchmove`
 - added `outview()` component to control behavior when object leaves visible area
 - deprecated `cleanup(delay?: number)` in favor of `cleanup(opt?: CleanupOpt)`
+- deprecated `mouseWorldPos()` in favor of `toWorld(mousePos())`
 
 ### v2000.1.8